xlat: Fix warning in CHECK_VIRT_ADDR_SPACE_SIZE
When TF is compiled for aarch32 MAX_VIRT_ADDR_SPACE_SIZE is 2^32 in some cases,
which makes the test (size) <= MAX_VIRT_ADDR_SPACE_SIZE a tautology because
uintptr_t is a 32 bit value. The cast remove the warning for clang.

+++ include/lib/xlat_tables/xlat_tables_arch.h
@@ -21,8 +21,8 @@
  * limits. Not that these limits are different for AArch32 and AArch64.
  */
 #define CHECK_VIRT_ADDR_SPACE_SIZE(size)			\
-	(((size) >= MIN_VIRT_ADDR_SPACE_SIZE) &&		\
-	((size) <= MAX_VIRT_ADDR_SPACE_SIZE) &&			\
+	(((unsigned long long)(size) >= MIN_VIRT_ADDR_SPACE_SIZE) &&		\
+	((unsigned long long)(size) <= MAX_VIRT_ADDR_SPACE_SIZE) &&		\
 	IS_POWER_OF_TWO(size))
 
 /*
